flask_app/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
redis==5.0.0
elasticsearch==8.9.0
python-keycloak==3.3.0